+++ frontend/src/views/Home.vue
@@ -1,5 +1,4 @@
 <script setup lang="ts">
-import TheFooter from '@/components/TheFooter.vue';
 import TheNavBar from '@/components/TheNavBar.vue';
 </script>
 <template>
