+++ frontend/src/views/family/heads/Head.vue
